modify compile bugs
code compile successful, but function not test

+++ usDecode.h
@@ -19,12 +19,12 @@
 extern "C" {
 #endif
 
-uint8_t usDecode_BasicMagicHandle(struct uStorPro_headInfo *proHeader);
-uint8_t usDecode_ReadHandle(uStorPro_fsOnDev *proHeader, uint8_t *readHeader, uint32_t buffSize);
-uint8_t usDecode_WriteHandle(uStorPro_fsOnDev *proHeader, uint8_t *writeHeader, uint32_t payloadSize);
-uint8_t usDecode_CreateHandle(uStorPro_fsOnDev *proHeader, uint8_t *creatPtr, uint32_t ptrLen);
-uint8_t usDecode_DeleteHandle(uStorPro_fsOnDev *proHeader, uint8_t *delPtr, uint32_t ptrLen);
-uint8_t usDecode_MoveHandle(uStorPro_fsOnDev *proHeader, uint8_t *MovPtr, uint32_t ptrLen);
+int32_t usDecode_BasicMagicHandle(struct uStorPro_headInfo *proHeader);
+int32_t usDecode_ReadHandle(uStorPro_fsOnDev *proHeader, uint8_t *readHeader, uint32_t buffSize);
+int32_t usDecode_WriteHandle(uStorPro_fsOnDev *proHeader, uint8_t *writeHeader, uint32_t payloadSize);
+int32_t usDecode_CreateHandle(uStorPro_fsOnDev *proHeader, uint8_t *creatPtr, uint32_t ptrLen);
+int32_t usDecode_DeleteHandle(uStorPro_fsOnDev *proHeader, uint8_t *delPtr, uint32_t ptrLen);
+int32_t usDecode_MoveHandle(uStorPro_fsOnDev *proHeader, uint8_t *MovPtr, uint32_t ptrLen);
